+++ apps/acf-extension/src/content_scripts/i18n.ts
@@ -1,0 +1,25 @@
+export const I18N_COMMON = {
+  REPEAT: chrome.i18n.getMessage('@REPEAT'),
+  DISABLED: chrome.i18n.getMessage('@DISABLED'),
+  GOTO: chrome.i18n.getMessage('@GOTO'),
+  RECHECK: chrome.i18n.getMessage('@RECHECK'),
+  COMPARE: chrome.i18n.getMessage('@COMPARE'),
+  RESULT: chrome.i18n.getMessage('@RESULT'),
+  RECHECK_OPTION: chrome.i18n.getMessage('@RECHECK_OPTION'),
+  DEFAULT: chrome.i18n.getMessage('@DEFAULT'),
+  COMPLETED: chrome.i18n.getMessage('@COMPLETED'),
+  ERROR: chrome.i18n.getMessage('@ERROR'),
+  SCHEDULE: chrome.i18n.getMessage('@SCHEDULE'),
+};
+
+export const I18N_ERROR = {
+  ELEMENT_FINDER_BLANK: chrome.i18n.getMessage('@ERROR__ELEMENT_FINDER_BLANK'),
+  NO_MATCH: chrome.i18n.getMessage('@ERROR__NO_MATCH'),
+  WRONG_TITLE: chrome.i18n.getMessage('@ERROR__WRONG_TITLE'),
+  WRONG_DESCRIPTION: chrome.i18n.getMessage('@ERROR__WRONG_DESCRIPTION'),
+  ELEMENT_FINDER_TITLE: chrome.i18n.getMessage('@ERROR__ELEMENT_FINDER_TITLE'),
+  ELEMENT_FINDER_DESCRIPTION: chrome.i18n.getMessage('@ERROR__ELEMENT_FINDER_DESCRIPTION'),
+  INVALID_ELEMENT_FINDER: chrome.i18n.getMessage('@ERROR__INVALID_ELEMENT_FINDER'),
+  NOT_FOUND_STOP: chrome.i18n.getMessage('@ERROR__NOT_FOUND_STOP'),
+  NOT_FOUND_RELOAD: chrome.i18n.getMessage('@ERROR__NOT_FOUND_RELOAD'),
+};