+++ apps/acf-options-page/src/App.tsx
@@ -2,7 +2,6 @@ import { Suspense, useEffect } from 'react';
 import Header from './app/header';
 import { ToastHandler, DataList, Loading } from './components';
 import { BlogModal, ExtensionNotFoundModal } from './modal';
-import { APP_NAME } from './constants';
 import { useAppDispatch, useAppSelector } from './hooks';
 import { appSelector, getManifest } from './store/app.slice';
 import ConfirmationModalContextProvider from './_providers/confirm.provider';
@@ -14,14 +13,6 @@ function App() {
   useEffect(() => {
     dispatch(getManifest());
   }, [dispatch]);
-
-  useEffect(() => {
-    if (/(DEV|BETA|LOCAL)/.test(process.env.NX_VARIANT || '')) {
-      window.document.title = `${APP_NAME} [${process.env.NX_VARIANT}]`;
-    } else {
-      window.document.title = APP_NAME;
-    }
-  }, []);
 
   return (
     <Suspense fallback={<Loading message='Loading localization...' className='m-5 p-5' />}>