+++ apps/acf-options-page/src/store/config/config.middleware.ts
@@ -1,5 +1,4 @@
 import { AnyAction, createListenerMiddleware, isAnyOf } from '@reduxjs/toolkit';
-import { getI18n } from 'react-i18next';
 import {
   addConfig,
   duplicateConfig,
@@ -34,18 +33,17 @@ import {
   setActionStatementError,
   setActionStatementMessage,
 } from './action';
+import i18next from 'i18next';
 
 const configsToastListenerMiddleware = createListenerMiddleware();
 configsToastListenerMiddleware.startListening({
   matcher: isAnyOf(addConfig, removeConfig, duplicateConfig),
   effect: async (action, listenerApi) => {
     const [type, method] = action.type.split('/');
-    const i18n = getI18n();
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const language: any = i18n.getDataByLanguage(i18n.language)?.['web-new'];
 
-    const header = language.toast[type][method]?.header?.replace('{{name}}', type);
-    const body = language.toast[type][method]?.body?.replace('{{name}}', type);
+    const header = i18next.t(`toast.${type}.${method}.header`, { name: type });
+    const body = i18next.t(`toast.${type}.${method}.body`, { name: type });
     if (header) {
       listenerApi.dispatch(addToast({ header, body }));
     }
@@ -53,24 +51,24 @@ configsToastListenerMiddleware.startListening({
 });
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-const getMessageFunc = (action: AnyAction, language): { success: any; failure: any; message: string } => {
+const getMessageFunc = (action: AnyAction): { success: any; failure: any; message: string } => {
   switch (action.type) {
     case updateConfigSettings.type:
-      return { success: setConfigSettingsMessage, failure: setConfigSettingsError, message: language.message?.configSettings };
+      return { success: setConfigSettingsMessage, failure: setConfigSettingsError, message: i18next.t(`message.configSettings`) };
     case updateBatch.type:
-      return { success: setBatchMessage, failure: setBatchError, message: language.message?.batch };
+      return { success: setBatchMessage, failure: setBatchError, message: i18next.t(`message.batch`) };
     case updateAction.type:
     case reorderActions.type:
     case removeAction.type:
-      return { success: setActionMessage, failure: setActionError, message: language.message?.action };
+      return { success: setActionMessage, failure: setActionError, message: i18next.t(`message.action`) };
     case syncActionAddon.type:
-      return { success: setActionAddonMessage, failure: setActionAddonError, message: language.message?.actionAddon };
+      return { success: setActionAddonMessage, failure: setActionAddonError, message: i18next.t(`message.actionAddon`) };
     case syncActionStatement.type:
-      return { success: setActionStatementMessage, failure: setActionStatementError, message: language.message?.actionStatement };
+      return { success: setActionStatementMessage, failure: setActionStatementError, message: i18next.t(`message.actionStatement`) };
     case syncActionSettings.type:
-      return { success: setActionSettingsMessage, failure: setActionSettingsError, message: language.message?.actionSettings };
+      return { success: setActionSettingsMessage, failure: setActionSettingsError, message: i18next.t(`message.actionSettings`) };
     default:
-      return { success: setConfigMessage, failure: setConfigError, message: language.message?.config };
+      return { success: setConfigMessage, failure: setConfigError, message: i18next.t(`message.config`) };
   }
 };
 
@@ -94,18 +92,18 @@ configsListenerMiddleware.startListening({
   effect: async (action, listenerApi) => {
     // Run whatever additional side-effect-y logic you want here
     const state = listenerApi.getState() as RootState;
-    const i18n = getI18n();
+
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const language: any = i18n.getDataByLanguage(i18n.language)?.['web-new'];
-    const { success, failure, message } = getMessageFunc(action, language);
 
     StorageService.set(window.EXTENSION_ID, { [LOCAL_STORAGE_KEY.CONFIGS]: state.configuration.configs })
       .then(() => {
+        const { success, message } = getMessageFunc(action);
         if (success && message) {
           listenerApi.dispatch(success(message));
         }
       })
       .catch((error) => {
+        const { failure } = getMessageFunc(action);
         if (failure) {
           if (error instanceof Error) {
             listenerApi.dispatch(failure(error.message));