+++ libs/acf/common/src/lib/model/index.ts
@@ -3,3 +3,7 @@ export * from './addon-model';
 export * from './batch-model';
 export * from './config-model';
 export * from './setting-model';
+
+export const BATCH_REPEAT = '<batchRepeat>';
+export const ACTION_REPEAT = '<actionRepeat>';
+export const SESSION_COUNT = '<sessionCount>';