+++ libs/acf/events/src/lib/attribute.events.ts
@@ -1,4 +1,4 @@
-import { Logger } from '@dhruv-techapps/core-common';
+import { ACTION_I18N_TITLE } from '.';
 import CommonEvents from './common.events';
 
 export const AttributeEvents = (() => {
@@ -12,7 +12,7 @@ export const AttributeEvents = (() => {
   };
 
   const start = (elements: Array<HTMLElement>, value: string) => {
-    Logger.colorDebug(`AttributeEvents`, value);
+    console.debug(`${ACTION_I18N_TITLE} #${window.__currentAction}`, elements, value);
     CommonEvents.loopElements(elements, value, execCommand);
   };
   return { start };