+++ libs/acf/events/src/lib/key.events.ts
@@ -1,10 +1,7 @@
-import { Logger, ConfigError, SystemError } from '@dhruv-techapps/core-common';
-import {} from '@dhruv-techapps/acf-common';
-import CommonEvents, { UNKNOWN_ELEMENT_TYPE_ERROR } from './common.events';
+import { ConfigError, SystemError } from '@dhruv-techapps/core-common';
 import { Timer } from '@dhruv-techapps/shared-util';
-
-// KeyEvents::{value:'Example text',delay:300}
-// KeyEvents::Example text
+import { ACTION_I18N_TITLE } from '.';
+import CommonEvents, { UNKNOWN_ELEMENT_TYPE_ERROR } from './common.events';
 
 const KEYBOARD_EVENT_KEYDOWN = 'keydown';
 const KEYBOARD_EVENT_KEYUP = 'keyup';
@@ -16,7 +13,6 @@ type KeyEvent = KeyboardEventInit & {
 
 export const KeyEvents = (() => {
   const getVerifiedEvents = (events: string): Array<KeyEvent> => {
-    Logger.colorDebug(`getVerifiedEvents`, events);
     if (!events) {
       throw new SystemError('Event is blank!', 'Event cant be blank | null | undefined');
     }
@@ -31,11 +27,6 @@ export const KeyEvents = (() => {
         throw new ConfigError(events, 'Invalid Events');
       }
     } catch (error) {
-      if (error instanceof Error) {
-        Logger.colorError(error.message);
-      } else {
-        Logger.colorError(JSON.stringify(error));
-      }
       result = events.split('').map((event) => ({ ...CommonEvents.getKeyboardEventProperties({ key: event }) }));
     }
     if (result) {
@@ -63,7 +54,7 @@ export const KeyEvents = (() => {
 
   const start = (elements: Array<HTMLElement>, event: string) => {
     const events = getVerifiedEvents(event);
-    Logger.colorDebug(`KeyEvents`, events);
+    console.debug(`${ACTION_I18N_TITLE} #${window.__currentAction}`, elements, events);
     CommonEvents.loopElements<HTMLElement, Array<KeyEvent>>(elements, events, dispatchEvent);
   };
   return { start };