+++ libs/acf/events/src/lib/keyboard.events.ts
@@ -1,4 +1,5 @@
-import { Logger, SystemError } from '@dhruv-techapps/core-common';
+import { SystemError } from '@dhruv-techapps/core-common';
+import { ACTION_I18N_TITLE } from '.';
 import CommonEvents from './common.events';
 
 const KEYBOARD_EVENT_KEYDOWN = 'keydown';
@@ -7,7 +8,6 @@ const KEYBOARD_EVENT_KEYPRESS = 'keypress';
 
 export const KeyboardEvents = (() => {
   const getVerifiedEvents = (events: string): KeyboardEventInit => {
-    Logger.colorDebug(`getVerifiedEvents`, events);
     if (!events) {
       throw new SystemError('Event is blank!', 'Event cant be blank | null | undefined');
     }
@@ -51,7 +51,7 @@ export const KeyboardEvents = (() => {
 
   const start = (elements: Array<HTMLElement>, event: string) => {
     const events = getVerifiedEvents(event);
-    Logger.colorDebug(`Keyboard Events`, events);
+    console.debug(`${ACTION_I18N_TITLE} #${window.__currentAction}`, elements, events);
     CommonEvents.loopElements(elements, events, dispatchEvent);
   };
   return { start };