+++ libs/acf/events/src/lib/mouse.events.ts
@@ -1,4 +1,5 @@
-import { Logger, ConfigError, SystemError } from '@dhruv-techapps/core-common';
+import { ConfigError, SystemError } from '@dhruv-techapps/core-common';
+import { ACTION_I18N_TITLE } from '.';
 import CommonEvents from './common.events';
 
 type CustomMouseEvent = {
@@ -17,7 +18,6 @@ export const MouseEvents = (() => {
   };
 
   const getVerifiedEvents = (events: string): Array<string | CustomMouseEvent> => {
-    Logger.colorDebug('GetVerifiedEvents', events);
     if (!events) {
       throw new SystemError('Event is blank!', 'Event cant be blank | null | undefined');
     }
@@ -46,7 +46,7 @@ export const MouseEvents = (() => {
 
   const start = (elements: Array<HTMLElement>, event: string) => {
     const events = getVerifiedEvents(event);
-    Logger.colorDebug(`MouseEvents`, events);
+    console.debug(`${ACTION_I18N_TITLE} #${window.__currentAction}`, elements, events);
     CommonEvents.loopElements(elements, events, dispatchEvent);
   };
   return { start };