+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 3.12.0 (2024-05-13)
+
+### 🚀 Features
+
+- Add version check for tour in Header component ([#277](https://github.com/Dhruv-Techapps/auto-clicker-auto-fill/pull/277))
+- Update NX_RELEASE_VERSION to 9.9.9 in acf-extension and acf-options-page ([#280](https://github.com/Dhruv-Techapps/auto-clicker-auto-fill/pull/280))
+- Update language translations for Korean, Chinese, and French ([ef339ce](https://github.com/Dhruv-Techapps/auto-clicker-auto-fill/commit/ef339ce))
+- Update context menu titles for field configuration, auto clicker, and configuration page ([cd13cb3](https://github.com/Dhruv-Techapps/auto-clicker-auto-fill/commit/cd13cb3))
+- Update context menu titles for field configuration, auto clicker, and configuration page ([28c4ef4](https://github.com/Dhruv-Techapps/auto-clicker-auto-fill/commit/28c4ef4))
+- Update context menu titles for field configuration, auto clicker, and configuration page ([d430b3b](https://github.com/Dhruv-Techapps/auto-clicker-auto-fill/commit/d430b3b))
+- Update context menu titles for field configuration, auto clicker, and configuration page ([525bda8](https://github.com/Dhruv-Techapps/auto-clicker-auto-fill/commit/525bda8))
+- Update context menu titles for field configuration, auto clicker, and configuration page ([bede29d](https://github.com/Dhruv-Techapps/auto-clicker-auto-fill/commit/bede29d))
+- Update language translations for Korean, Chinese, and French ([97cc15c](https://github.com/Dhruv-Techapps/auto-clicker-auto-fill/commit/97cc15c))
+
+### ❤️ Thank You
+
+- Dharmesh @dharmesh-hemaram
+- Dharmesh Patel @dharmesh-hemaram
+
 ## 4.0.2 (2024-05-11)
 
 This was a version bump only, there were no code changes.